pyproject.toml
[tool.poetry]
name = "testfastapi"
version = "0.1.0"
description = ""
authors = ["JustDreamer7 <ad_pochestnev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
faker = "^19.3.1"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.4.2"}
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
